mkdir and check file gpg exist

+++ install.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# Pedir al usuario los valores de las variables necesarias
+read -p "Ingrese el subdominio de su ISPbrain: " subdomain
+read -p "Ingrese el id de la cuenta de AltWha: " account
+read -p "Ingrese el usuario de API ISPbrain: " user
+read -p "Ingrese la contraseña de API ISPbrain: " password
+
 # Instalar docker
 apt-get update
 apt-get install -y \
@@ -7,9 +13,11 @@ apt-get install -y \
     curl \
     gnupg
 
-install -m 0755 -d /etc/apt/keyrings
-curl -fsSL https://download.docker.com/linux/debian/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-chmod a+r /etc/apt/keyrings/docker.gpg
+if ! test -f /etc/apt/keyrings/docker.gpg; then
+    install -m 0755 -d /etc/apt/keyrings
+    curl -fsSL https://download.docker.com/linux/debian/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+    chmod a+r /etc/apt/keyrings/docker.gpg
+fi
 
 echo \
   "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
@@ -20,11 +28,7 @@ apt-get update
 
 apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
 
-# Pedir al usuario los valores de las variables necesarias
-read -p "Ingrese el subdominio de su ISPbrain: " subdomain
-read -p "Ingrese el id de la cuenta de AltWha: " account
-read -p "Ingrese el usuario de API ISPbrain: " user
-read -p "Ingrese la contraseña de API ISPbrain: " password
+mkdir /root/log
 
 # Correr el docker con las variables ingresadas por el usuario
 docker run -d \
@@ -32,5 +36,6 @@ docker run -d \
 --env ISPBRAIN_ACCOUNT="$account" \
 --env ISPBRAIN_USER="$user" \
 --env ISPBRAIN_PASSWORD="$password" \
+--restart=always \
 -v /root/.local/share/mudslide:/usr/src/app/cache \
--v log:/root/AltWha/log crenein/altwhasender:v1.0.3
+-v /root/log:/root/AltWha/log crenein/altwhasender:v1.0.6
